FEM: geom tools, move find element in shape and is same geometry from meshtools
